Refaktorointia ja testejä
