--- tests/data/data.js
@@ -1,15 +1,0 @@
-const TKT10001 = {
-    "code": "TKT10001",
-    "name": "Johdatus tietojenkäsittelytieteeseen",
-    "level": "Perusopinnot",
-    "compulsory": true,
-    "prereqs": [],
-    "studytrack": [],
-    "ects": "5",
-    "url": "https://courses.helsinki.fi/fi/TKT10001",
-    "visible": true
-}
-
-module.exports = {
-    TKT10001
-}
