+++ tests/data/dummyCourseData.js
@@ -1,0 +1,27 @@
+const TKT10001 = {
+    "code": "TKT10001",
+    "name": "Johdatus tietojenkäsittelytieteeseen",
+    "level": "Perusopinnot",
+    "compulsory": true,
+    "prereqs": [],
+    "studytrack": [],
+    "ects": "5",
+    "url": "https://courses.helsinki.fi/fi/TKT10001"
+}
+
+const CSM14209 = {
+    "code": "CSM14209",
+    "name": "Seminar in Novel Software Architecture Design",
+    "level": "Syventävät",
+    "compulsory": false,
+    "prereqs": [],
+    "studytrack": [
+        "ss"
+    ],
+    "ects": "5",
+    "url": "https://courses.helsinki.fi/fi/CSM14209"
+}
+
+module.exports = {
+    TKT10001, CSM14209
+}
